Download spatial hexagonal grid H3

Results of the AOP project are spatially aggregated on a hexagonal grid based on the global H3 index at resolution 8, with a size of 357 meters (short diagonal) and an area of 0.74 km2. More information about H3 at https://h3geo.org/docs/core-library/restable/. See documentation 'Details' for the data dictionary.

read_grid(city, showProgress = TRUE)

Arguments

city
Character. A city name or three-letter abbreviation. If city="all", results for all cities are loaded.

showProgress
Logical. Defaults to TRUE display progress bar

Value

An sf data.frame object

Data dictionary:

Data type	column	Description

geographic	abbrev_muni	Abbreviation of city name (3 letters)

geographic	name_muni	City name

geographic	code_muni	7-digit code of each city

geographic	id_hex	Unique id of hexagonal cell

Examples

# \donttest{
# Read spatial grid of a single city
nat <- read_grid(city = 'Natal', showProgress = FALSE)

#> Downloading: 780 B     Downloading: 780 B     Downloading: 1.2 kB     Downloading: 1.2 kB     Downloading: 1.2 kB     Downloading: 1.2 kB     Downloading: 1.6 kB     Downloading: 1.6 kB     Downloading: 1.6 kB     Downloading: 1.6 kB     Downloading: 1.7 kB     Downloading: 1.7 kB     Downloading: 1.8 kB     Downloading: 1.8 kB     Downloading: 1.8 kB     Downloading: 1.8 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2 kB     Downloading: 2.7 kB     Downloading: 2.7 kB     Downloading: 2.7 kB     Downloading: 2.7 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 11 kB     Downloading: 19 kB     Downloading: 19 kB     Downloading: 19 kB     Downloading: 19 kB     Downloading: 19 kB     Downloading: 19 kB     Downloading: 27 kB     Downloading: 27 kB     Downloading: 27 kB     Downloading: 27 kB     Downloading: 27 kB     Downloading: 27 kB     Downloading: 35 kB     Downloading: 35 kB     Downloading: 35 kB     Downloading: 35 kB     Downloading: 35 kB     Downloading: 35 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 51 kB     Downloading: 51 kB     Downloading: 51 kB     Downloading: 51 kB     Downloading: 51 kB     Downloading: 51 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 76 kB     Downloading: 76 kB     Downloading: 76 kB     Downloading: 76 kB     Downloading: 76 kB     Downloading: 76 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 84 kB     Downloading: 92 kB     Downloading: 92 kB     Downloading: 92 kB     Downloading: 92 kB     Downloading: 92 kB     Downloading: 92 kB     Downloading: 98 kB     Downloading: 98 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 120 kB     Downloading: 120 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     

# Read spatial grid of all cities in the project
# all <- read_grid(city = 'all')
# }