Download accessibility estimates with population and land use data

Download annual estimates of access to employment, health and education services by transport mode and time of the day. See documentation 'Details' for the data dictionary. The data comes aggregated on a hexagonal grid based on the global H3 index at resolution 9, with a size of 357 meters (short diagonal) and an area of 0.74 km2. More information about H3 at https://h3geo.org/docs/core-library/restable/.

read_access(
  city,
  mode = "walk",
  peak = TRUE,
  year = 2019,
  geometry = FALSE,
  showProgress = TRUE
)

Arguments

city
Character. A city name or three-letter abbreviation. If city="all", results for all cities are loaded.

mode
Character. A transport mode. Modes available include 'public_transport', 'bicycle', or 'walk' (the default).

peak
Logical. If TRUE (the default), returns accessibility estimates during peak time, between 6am and 8am. If FALSE, returns accessibility during off-peak, between 2pm and 4am. This argument only takes effect when mode = public_transport.

year
Numeric. A year number in YYYY format. Default set to 2019, the only year currently available.

geometry
Logical. If FALSE (the default), returns a regular data.table of aop data. If TRUE, returns a an sf data.frame with simple feature geometry of spatial hexagonal grid H3. See details in read_grid.

showProgress
Logical. Defaults to TRUE display progress bar

Value

A data.frame object

Data dictionary:

The name of the columns with accessibility estimates are the junction of three components: 1) Indicator 2) Type of opportunity 3) Time thresold (if applicable)

1) Indicator

Indicator	Description	Note

CMA	Cumulative opportunity measure (active)

TMI	Travel time to closest opportunity	Value = Inf when travel time is longer than 2h (public transport) or 1,5h (walking or bicycle)

2) Type of opportunity

Indicator	Description	Note

TT	All jobs

TQ	Total jobs with partial match between job education and income quintile

TD	Total jobs with partial match between job education and income decile

ST	All healthcare facilities

SB	Healthcare facilities - Low complexity

SM	Healthcare facilities - Medium complexity

SA	Healthcare facilities - High complexity

ET	All public schools

EI	Public schools - early childhood

EF	Public schools - elementary schools

EM	Public schools - high schools

3) Time thresold (only applicable to CMA estimates)y

Time thresold	Description	Note - Only applicable to:

15	Opportunities accessible within 15 min.	Active transport modes

30	Opportunities accessible within 30 min.	All transport modes

45	Opportunities accessible within 45 min.	Active transport modes

60	Opportunities accessible within 60 min.	All transport modes

90	Opportunities accessible within 90 min.	Public transport

120	Opportunities accessible within 120 min.	Public transport

Examples

# \donttest{
# Read accessibility estimates of a single city
df <- read_access(city = 'Fortaleza', mode = 'public_transport', year = 2019, showProgress = FALSE)
#> Using year 2019
#> Using mode public_transport
#> Using year 2019
#> Using year 2010
df <- read_access(city = 'for', mode = 'public_transport', year = 2019, showProgress = FALSE)
#> Using year 2019
#> Using mode public_transport
#> Using year 2019
#> Using year 2010

# Read accessibility estimates for all cities
all <- read_access(city = 'all', mode = 'walk', year = 2019, showProgress = FALSE)
#> Using year 2019
#> Using mode walk
#> Using year 2019
#> Using year 2010
# }