Download spatial hexagonal grid H3

Results of the AOP project are spatially aggregated on a hexagonal grid-based on the global H3 index at resolution 9, with a size of 357 meters (short diagonal) and an area of 0.74 km2. More information about H3 at https://h3geo.org/docs/core-library/restable/. See documentation 'Details' for the data dictionary.

read_grid(city = NULL, showProgress = FALSE)

Arguments

city

Character. A city name or three-letter abbreviation. If city="all", results for all cities are loaded.

showProgress

Logical. Defaults to TRUE display progress bar

Value

An sf data.frame object

Data dictionary:

Data type	column	Description

geographic	abbrev_muni	Abbreviation of city name (3 letters)

geographic	name_muni	City name

geographic	code_muni	7-digit code of each city

geographic	id_hex	Unique id of hexagonal cell

Cities available

City name	Three-letter abbreviation

Belem	bel

Belo Horizonte	bho

Brasilia	bsb

Campinas	cam

Campo Grande	cgr

Curitiba	cur

Duque de Caxias	duq

Fortaleza	for

Goiania	goi

Guarulhos	gua

Maceio	mac

Manaus	man

Natal	nat

Porto Alegre	poa

Recife	rec

Rio de Janeiro	rio

Salvador	sal

Sao Goncalo	sgo

Sao Luis	slz

Sao Paulo	spo

Examples

if (FALSE)  if (interactive()) {
# Read spatial grid of a single city
nat <- read_grid(city = 'Natal', showProgress = FALSE)

# Read spatial grid of all cities in the project
# all <- read_grid(city = 'all', showProgress = FALSE)
}