Download spatial hexagonal grid H3

Results of the AOP project are spatially aggregated on a hexagonal grid based on the global H3 index at resolution 8, with a size of 357 meters (short diagonal) and an area of 0.74 km<sup>2</sup>. More infor about H3 at https://h3geo.org/docs/core-library/restable.

read_grid(city, showProgress = TRUE)

Arguments

city
Character. A city name or three-letter abbreviation. If city="all", results for all cities are loaded.

showProgress
Logical. Defaults to TRUE display progress bar

Value

An sf data.frame object

Examples

# \donttest{
# Read spatial grid of a single city
nat <- read_grid(city = 'natal')

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Downloading: 1.5 kB     Downloading: 1.5 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 1.9 kB     Downloading: 18 kB     Downloading: 18 kB     Downloading: 34 kB     Downloading: 34 kB     Downloading: 50 kB     Downloading: 50 kB     Downloading: 50 kB     Downloading: 50 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 75 kB     Downloading: 75 kB     Downloading: 91 kB     Downloading: 91 kB     Downloading: 100 kB     Downloading: 100 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     

nat <- read_grid(city = 'nat')

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Downloading: 780 B     Downloading: 780 B     Downloading: 1.5 kB     Downloading: 1.5 kB     Downloading: 2.1 kB     Downloading: 2.1 kB     Downloading: 2.1 kB     Downloading: 2.1 kB     Downloading: 18 kB     Downloading: 18 kB     Downloading: 34 kB     Downloading: 34 kB     Downloading: 43 kB     Downloading: 43 kB     Downloading: 59 kB     Downloading: 59 kB     Downloading: 67 kB     Downloading: 67 kB     Downloading: 83 kB     Downloading: 83 kB     Downloading: 99 kB     Downloading: 99 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 110 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 130 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     Downloading: 140 kB     


# Read spatial grid of all cities in the project
all <- read_grid(city = 'all')

#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |==========                                                            |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |========================                                              |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |====================================================                  |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |==================================================================    |  95%  |                                                                              |======================================================================| 100%

# }